# TimerOutputs + some speedups