--- a/Project.toml
+++ b/Project.toml
@@ -8,8 +8,10 @@ Distributed = "8ba89e20-285c-5b6f-9357-94700520ee1b"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 
 [compat]
-StatsBase = "0.33"
 Distributions = "0.24"
+StatsBase = "0.33"
+TimerOutputs = "0.5"
 julia = "1"