--- a/src/tmcmc.jl
+++ b/src/tmcmc.jl
@@ -21,6 +21,9 @@
 #
 ###
 
+using TimerOutputs
+const to = TimerOutput()
+
 function tmcmc(
     log_fD_T::Function,
     log_fT::Function,
@@ -39,11 +42,13 @@ function tmcmc(
 
     Ndims = size(θ_j, 2)         # Number of dimensions (input)
 
-    while βj < 1
+    Σ_j = zeros(Ndims, Ndims)
+
+    @timeit_debug to "Main while loop" while βj < 1
 
         j1 = j1 + 1
 
-        @debug "Beginnig iteration $j1"
+        @debug "Beginning iteration $j1"
 
         ###
         # Parallel evaluation of the likelihood
@@ -51,7 +56,7 @@ function tmcmc(
 
         @debug "Computing likelihood with $(nworkers()) workers..."
 
-        ins = [θ_j[i,:] for i in 1:size(θ_j, 1)]
+        @timeit_debug to "Initialize ins" ins = [θ_j[i,:] for i in 1:size(θ_j, 1)]
         Lp_j = pmap(log_fD_T, ins)
         Lp_j = reduce(vcat, Lp_j)
 
@@ -64,7 +69,7 @@ function tmcmc(
         low_β = βj; hi_β = 2; Lp_adjust = maximum(Lp_j);
         x1 = (hi_β + low_β) / 2;
 
-        while (hi_β - low_β) / ((hi_β + low_β) / 2) > 1e-6
+        @timeit_debug to "Inner while loop" while (hi_β - low_β) / ((hi_β + low_β) / 2) > 1e-6
             x1 = (hi_β + low_β) / 2;
             wj_test = exp.((x1 .- βj ) .* (Lp_j .- Lp_adjust));
             cov_w   = std(wj_test) / mean(wj_test);
@@ -80,39 +85,44 @@ function tmcmc(
         ###
         @debug "Computing weights..."
 
-        w_j = exp.((βj1 - βj) .* (Lp_j .- Lp_adjust))       # Nominal weights from likilhood and βjs
+        @timeit_debug to "Nominal weights" w_j = exp.((βj1 - βj) .* (Lp_j .- Lp_adjust))       # Nominal weights from likilhood and βjs
 
-        Log_ev = log(mean(w_j)) + (βj1 - βj) * Lp_adjust + Log_ev   # Log evidence in current iteration
+        @timeit_debug to "Log evidence" Log_ev = log(mean(w_j)) + (βj1 - βj) * Lp_adjust + Log_ev   # Log evidence in current iteration
 
         # Normalised weights
-        wn_j = w_j ./ sum(w_j);
+        @timeit_debug to "Normalised weights" wn_j = w_j ./ sum(w_j);
 
-        Th_wm = θ_j .* wn_j                 # Weighted mean of samples
+        @timeit_debug to "Weighted mean" Th_wm = θ_j .* wn_j                 # Weighted mean of samples
 
-        ###
-        #   Calculation of COV matrix of proposal
-        ###
-        Σ_j    = zeros(Ndims, Ndims)
+        @timeit_debug to "Compute covariance" begin
+            ###
+            #   Calculation of COV matrix of proposal
+            ###
+            @timeit_debug to "Inititialize Σ_j" Σ_j .= 0
 
-        for l = 1:Nsamples
-            Σ_j = Σ_j + beta2 .* wn_j[l] .* (θ_j[l,:]' .- Th_wm)' * (θ_j[l,:]' .- Th_wm)
+            @timeit_debug to "Update Σ_j" for l = 1:Nsamples
+                Σ_j .+= beta2 .* wn_j[l] .* (θ_j[l,:]' .- Th_wm)' * (θ_j[l,:]' .- Th_wm)
+            end
+
+            # Ensure that cov is symetric
+            @timeit_debug to "Symmetrize Σ_j" begin
+                Σ_j .+= Σ_j'
+                Σ_j ./= 2
+            end
+
         end
-
-        # Ensure that cov is symetric
-        Σ_j = (Σ_j' + Σ_j) / 2
-
         prop = mu -> proprnd(mu, Σ_j, log_fT) # Anonymous function for proposal
 
         target = x -> log_fD_T(x) .* βj1 .+ log_fT(x) # Anonymous function for transitional distribution
 
         # Weighted resampling of θj (indecies with replacement)
-        randIndex = sample(1:Nsamples, Weights(wn_j), Nsamples, replace=true)
+        @timeit_debug to "Compute randIndex" randIndex = sample(1:Nsamples, Weights(wn_j), Nsamples, replace=true)
 
-        ins = [θ_j[randIndex[i], :] for i = 1:Nsamples]
+        @timeit_debug to "Update ins" ins = [θ_j[randIndex[i], :] for i = 1:Nsamples]
 
         @debug "Markov chains with $(nworkers()) workers..."
         # pmap is a parallel map
-        θ_j1 = pmap(x -> run_chains(target, prop, x, burnin, thin), ins)
+        @timeit_debug to "Run chains" θ_j1 = pmap(x -> run_chains(target, prop, x, burnin, thin), ins)
 
         θ_j1 = reduce(vcat, θ_j1)