@AnderGray 2021-05-02
Great stuff Chad! Thanks for this PR. I had no idea Julia prefers columns.

I think they should be merged as is, what do you think @FriesischScott 

Some of the changes from [PR26](https://github.com/AnderGray/TransitionalMCMC.jl/pull/26) should also improve allocation times.