@FriesischScott 2021-05-02
I agree. Good stuff.